Update GA workflow jobs for ubuntu24.04 (#415)
* use clang-tidy installed in ubuntu-24.04 runner image

* updated clang-format version to 18.1.3 (same as ubuntu-24.04)

* updated runner image tags for CI jobs in workflows/ubuntu.yml

* updated CONTRIBUTING.md regarding clang-format tool version

* fixed undefined reference link error and coverage loss

* [bot] run clang-format & amalgamation

* fixed coverage loss again

+++ include/fkYAML/detail/types/node_t.hpp
@@ -10,7 +10,6 @@
 #define FK_YAML_DETAIL_TYPES_NODE_T_HPP
 
 #include <cstdint>
-#include <string>
 
 #include <fkYAML/detail/macros/version_macros.hpp>
 #include <fkYAML/node_type.hpp>
@@ -18,7 +17,7 @@
 FK_YAML_DETAIL_NAMESPACE_BEGIN
 
 /// @brief Definition of node value types.
-enum class node_t : std::uint32_t {
+enum class node_t : std::uint8_t {
     SEQUENCE,     //!< sequence value type
     MAPPING,      //!< mapping value type
     NULL_OBJECT,  //!< null value type
